+++ packages/cloud-cognitive/src/components/TagSet/TagSet.test.js
@@ -15,9 +15,9 @@ import { TagSetModal } from './TagSetModal';
 import { mockHTMLElement } from '../../global/js/utils/test-helper';
 import uuidv4 from '../../global/js/utils/uuidv4';
 
-const { devtoolsAttribute, getDevtoolsId, prefix } = pkg;
+const { prefix } = pkg;
 
-const blockClass = `${prefix}--tag-set`;
+const blockClass = `${pkg.prefix}--tag-set`;
 const blockClassOverflow = `${prefix}--tag-set-overflow`;
 
 const tagLabel = (index) => `Tag ${index + 1}`;
@@ -227,9 +227,8 @@ describe(TagSet.displayName, () => {
   it('adds the Devtools attribute to the containing node', () => {
     render(<TagSet data-testid={dataTestId} />);
 
-    expect(screen.getByTestId(dataTestId)).toHaveAttribute(
-      devtoolsAttribute,
-      getDevtoolsId(TagSet.displayName)
+    expect(screen.getByTestId(dataTestId)).toHaveDevtoolsAttribute(
+      TagSet.displayName
     );
   });